Update completion facts service to use providers.
- Completions will be coming from multiple different sources in the future and in order to build a system that works in both VSCode, VSWindows and VSMac we need the ability to toggle certain types of completions on and off easily.
- Created a `DirectiveCompletionItemProvider` this was basically a direct copy from the facts service.
- Also moved tests from the facts service over to the new directive completion item provider.
- Made `RazorCompletionItem`s `IEquatable` to ease in testing now and in the future.

aspnet/AspNetCore#6364
