+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Completion/RazorCompletionItem.cs
@@ -2,10 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
+using Microsoft.Extensions.Internal;
 
 namespace Microsoft.CodeAnalysis.Razor.Completion
 {
-    internal sealed class RazorCompletionItem
+    internal sealed class RazorCompletionItem : IEquatable<RazorCompletionItem>
     {
         public RazorCompletionItem(
             string displayText, 
@@ -41,5 +42,51 @@ namespace Microsoft.CodeAnalysis.Razor.Completion
         public string Description { get; }
 
         public RazorCompletionItemKind Kind { get; }
+
+        public override bool Equals(object obj)
+        {
+            return Equals(obj as RazorCompletionItem);
+        }
+
+        public bool Equals(RazorCompletionItem other)
+        {
+            if (other == null)
+            {
+                return false;
+            }
+
+            if (!string.Equals(DisplayText, other.DisplayText, StringComparison.Ordinal))
+            {
+                return false;
+            }
+
+            if (!string.Equals(InsertText, other.InsertText, StringComparison.Ordinal))
+            {
+                return false;
+            }
+
+            if (!string.Equals(Description, other.Description, StringComparison.Ordinal))
+            {
+                return false;
+            }
+
+            if (Kind != other.Kind)
+            {
+                return false;
+            }
+
+            return true;
+        }
+
+        public override int GetHashCode()
+        {
+            var hashCodeCombiner = HashCodeCombiner.Start();
+            hashCodeCombiner.Add(DisplayText);
+            hashCodeCombiner.Add(InsertText);
+            hashCodeCombiner.Add(Description);
+            hashCodeCombiner.Add(Kind);
+
+            return hashCodeCombiner.CombinedHash;
+        }
     }
 }
